Concluído chamadas e clientes

+++ app/Http/Controllers/CustomerController.php
@@ -2,16 +2,34 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Http\Requests\CustomerRequest;
 use App\Customer;
-use DataTables;
+use App\Http\Requests\CustomerRequest;
+use App\User;
+use Illuminate\Http\Request;
 
 class CustomerController extends Controller
 {
-    public function index()
+    public function __construct()
     {
-        $customers = Customer::all();
+        $users = User::all();
+        view()->share('users', $users);
+    }
+
+    public function index(Request $request)
+    {
+        $name = $request->name;
+        $fantasyName = $request->fantasy_name;
+        $docNumber = $request->doc_number;
+        $city = $request->city;
+        $customers = Customer::when($name, function ($q, $name) {
+            return $q->where('name', 'like', '%' . $name . '%');
+        })->when($fantasyName, function ($q, $fantasyName) {
+            return $q->where('fantasy_name', 'like', '%' . $fantasyName . '%');
+        })->when($docNumber, function ($q, $docNumber) {
+            return $q->where('doc_number', 'like', '%' . $docNumber . '%');
+        })->when($city, function ($q, $city) {
+            return $q->where('city', 'like', '%' . $city . '%');
+        })->paginate(10);
         return view('customer.index', compact('customers'));
     }
 
@@ -24,7 +42,7 @@ class CustomerController extends Controller
     {
         $request->commit();
         session()->flash('alert', ['type' => 'success', 'message' => 'Cliente cadastrado com sucesso!']);
-        return redirect()->route('customer.index');
+        return redirect()->back();
     }
 
     public function edit(Customer $customer)
@@ -37,23 +55,31 @@ class CustomerController extends Controller
     {
         $request->update($customer);
         session()->flash('alert', ['type' => 'success', 'message' => 'Cliente atualizado com sucesso!']);
-        return redirect()->route('customer.index');
+        return redirect()->back();
     }
 
     public function delete(Customer $customer)
     {
         $customer = Customer::findOrFail($customer->id);
+        if ($customer->calls->count()>0){
+            session()->flash('alert', [
+                'type' => 'error',
+                'message' => "O cliente $customer->name não pode ser excluído pois possuí lançamentos associados a ele!"
+            ]);
+            return redirect()->back();
+        }
         $customer->delete();
-        return redirect()->route('customer.index')->with(['success', 'Cliente excluído com sucesso']);
+        session()->flash('alert', ['type' => 'success', 'message' => 'Cliente excluído com sucesso!']);
+        return redirect()->back();
     }
 
-    public function search($query)
+    public function search(Request $request)
     {
-        $customer = Customer::where('name', 'like', '%' . $query .'%')
-            ->orWhere('fantasy_name', 'like', '%' . $query .'%')
-            ->orWhere('doc_number', 'like', '%' . $query .'%')
+        $query = $request->get('term');
+        $customer = Customer::where('name', 'like', '%' . $query . '%')
+            ->orWhere('fantasy_name', 'like', '%' . $query . '%')
+            ->orWhere('doc_number', 'like', '%' . $query . '%')
             ->get();
         return json_encode($customer);
     }
-
 }
